% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.table.R
\name{clean.table}
\alias{clean.table}
\title{Function to clean}
\usage{
clean.table(data, variable.number, max.levels.of.factor = 7)
}
\arguments{
\item{data}{describtion}

\item{variable.number}{a number}

\item{max.level.of.factor}{a number}
}
\value{
data.frame() with the airline
}
\description{
Function to clean
}
\author{
AxH
}
