% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun2.R
\name{fun2}
\alias{fun2}
\title{a title i think}
\usage{
fun2(p.vals, number.categories)
}
\arguments{
\item{p.vals}{stuf}

\item{number.categories}{stuf}
}
\value{
A matrix of posterior probabilities
}
\description{
a title i think
}
