\name{sohaRd-package}
\alias{sohaRd-package}
\alias{sohaRd}
\docType{package}
\title{
\packageTitle{sohaRd}
}
\description{
\packageDescription{sohaRd}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sohaRd}
\packageIndices{sohaRd}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sohaRd}

Maintainer: \packageMaintainer{sohaRd}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
