% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.table.R
\name{clean.table}
\alias{clean.table}
\title{Titlesaa}
\usage{
clean.table(data, variable.number, max.levels.of.factor = 7)
}
\arguments{
\item{data}{s}

\item{variable.number}{s}

\item{max.levels.of.factor}{s}
}
\value{
a
}
\description{
Titlesaa
}
