% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.table.R
\name{clean.table}
\alias{clean.table}
\title{aTitle}
\usage{
clean.table(data, variable.number, max.levels.of.factor = 7)
}
\arguments{
\item{data}{a}

\item{variable.number}{a}

\item{max.levels.of.factor}{a}
}
\value{
a
}
\description{
aTitle
}
