% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.table.R
\name{clean.table}
\alias{clean.table}
\title{Function searches a specific airline by IATA code}
\usage{
clean.table(data, variable.number, max.levels.of.factor = 7)
}
\arguments{
\item{data}{variable.number max.level.of.factor}
}
\value{
data.frame() with the airline
}
\description{
Function searches a specific airline by IATA code
}
\author{
who?
}
